refactor(uilanding): extract WebSocket handler with tests
Extract topoDataHandler from uilanding.py into handlers/websocket.py
to continue the module decomposition pattern established by other handler
extractions in this branch.

What changed:
- New handlers/websocket.py: topoDataHandler class with all WebSocket
  lifecycle methods (open, on_message, on_close), async keepalive loop,
  ping/pong latency measurement, CVP token refresh, connectivity event
  handling, debug mode toggle, and session summary logging
- Moved prune_recent_sessions() and RECONNECT_WINDOW_SECONDS constant
  into websocket.py (only used by this handler)
- Dependencies injected via initialize(): session_state dict, exam_state
  dict, and cvp_token_fn callable (avoids circular import)
- Updated uilanding.py: add import, remove class + helpers, update /td-ws
  route to pass _ws_kwargs, remove now-unused timedelta and
  tornado.websocket imports
- New tests/test_websocket.py: 11 tests covering session creation,
  session removal on close, cross-site origin rejection (security check),
  matching-host origin acceptance, _closed guard regression, and pong
  RTT update

All 282 tests pass (271 pre-existing + 11 new).
